riscv: support Allwinner D1 interrupt controllers
The main interrupt controller of the D1 is the PLIC, but it does not
provide any interface for configuring IRQ trigger type or wakeup.

These features are supported through registers in a separate memory
region (RISCV_CFG), and the device tree defines another interrupt
controller (INTC) as a wrapper around the PLIC and those registers.

Since most peripherals use INTC as their interrupt parent, and do not
use IRQ trigger type or wakeup features, do the following:
 - add the D1 compatible data to the PLIC driver,
 - add a stub driver for INTC to route things back to the PLIC.

Signed-off-by: Julien Cassette <julien.cassette@gmail.com>